docs(readme): add note about org visibility (#16)
By default members of an org are added with their "Organization
Visibility" set to Private, which will cause the action to return false,
as expected.  However, adding the note will save folks a lot of time as,
in case they consistently get wrong results, they at least have a hint
on what to check.

Fixes: #13

Signed-off-by: Fabiano Fidêncio <fabiano@fidencio.org>